Рефакторинг. Добавление XML комментарий.
