+++ SecurityDoors.RemoteControl/SecurityDoors.RemoteControl.csproj
@@ -12,4 +12,9 @@
     <ProjectReference Include="..\SecurityDoors.DataAccessLayer\SecurityDoors.DataAccessLayer.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Builders\" />
+    <Folder Include="cli\" />
+  </ItemGroup>
+
 </Project>
